Updated and fixed mvc test application
