+++ samples/ASPNet/MvcTestApplication.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvcTestApplication", "MvcTestApplication\MvcTestApplication.csproj", "{2658E893-EB33-43FE-B4D0-AD16BC76578B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2658E893-EB33-43FE-B4D0-AD16BC76578B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2658E893-EB33-43FE-B4D0-AD16BC76578B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2658E893-EB33-43FE-B4D0-AD16BC76578B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2658E893-EB33-43FE-B4D0-AD16BC76578B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9288D4A6-BBEC-4995-89CE-5E5AC37851B4}
+	EndGlobalSection
+EndGlobal
