+++ samples/ASPNet/MvcTestApplication/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="MvcTestApplication.MvcApplication" Language="C#" %>
