+++ samples/ASPNet/MvcTestApplication/IApplicationDependency.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace MvcTestApplication
+{
+    public interface IApplicationDependency
+    {
+        string GetContent();
+    }
+}
