+++ samples/ASPNet/MvcTestApplication/IRequestDependency.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace MvcTestApplication
+{
+    public interface IRequestDependency
+    {
+        string GetContent();
+    }
+}
