--- samples/ASPNet/MvcTestApplication/MvcTestApplication.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvcTestApplication", "MvcTestApplication\MvcTestApplication.csproj", "{16A55B8D-16C7-4EF6-9EDF-2320459E8990}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{16A55B8D-16C7-4EF6-9EDF-2320459E8990}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{16A55B8D-16C7-4EF6-9EDF-2320459E8990}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{16A55B8D-16C7-4EF6-9EDF-2320459E8990}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{16A55B8D-16C7-4EF6-9EDF-2320459E8990}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
