--- samples/ASPNet/MvcTestApplication/MvcTestApplication/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="MvcTestApplication.MvcApplication" Language="C#" %>
