--- samples/ASPNet/MvcTestApplication/MvcTestApplication/IApplicationDependency.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace MvcTestApplication
-{
-    public interface IApplicationDependency
-    {
-        string GetContent();
-    }
-}
