--- samples/ASPNet/MvcTestApplication/MvcTestApplication/IRequestDependency.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace MvcTestApplication
-{
-    public interface IRequestDependency
-    {
-        string GetContent();
-    }
-}
