+++ samples/ASPNet/MvcTestApplication/Views/Account/ChangePassword.aspx
@@ -1,0 +1,52 @@
+﻿<%@ Page Language="C#" MasterPageFile="~/Views/Shared/Site.Master" Inherits="System.Web.Mvc.ViewPage<MvcTestApplication.Models.ChangePasswordModel>" %>
+
+<asp:Content ID="changePasswordTitle" ContentPlaceHolderID="TitleContent" runat="server">
+    Change Password
+</asp:Content>
+
+<asp:Content ID="changePasswordContent" ContentPlaceHolderID="MainContent" runat="server">
+    <h2>Change Password</h2>
+    <p>
+        Use the form below to change your password. 
+    </p>
+    <p>
+        New passwords are required to be a minimum of <%: ViewData["PasswordLength"] %> characters in length.
+    </p>
+
+    <% using (Html.BeginForm()) { %>
+        <%: Html.ValidationSummary(true, "Password change was unsuccessful. Please correct the errors and try again.") %>
+        <div>
+            <fieldset>
+                <legend>Account Information</legend>
+                
+                <div class="editor-label">
+                    <%: Html.LabelFor(m => m.OldPassword) %>
+                </div>
+                <div class="editor-field">
+                    <%: Html.PasswordFor(m => m.OldPassword) %>
+                    <%: Html.ValidationMessageFor(m => m.OldPassword) %>
+                </div>
+                
+                <div class="editor-label">
+                    <%: Html.LabelFor(m => m.NewPassword) %>
+                </div>
+                <div class="editor-field">
+                    <%: Html.PasswordFor(m => m.NewPassword) %>
+                    <%: Html.ValidationMessageFor(m => m.NewPassword) %>
+                </div>
+                
+                <div class="editor-label">
+                    <%: Html.LabelFor(m => m.ConfirmPassword) %>
+                </div>
+                <div class="editor-field">
+                    <%: Html.PasswordFor(m => m.ConfirmPassword) %>
+                    <%: Html.ValidationMessageFor(m => m.ConfirmPassword) %>
+                </div>
+                
+                <p>
+                    <input type="submit" value="Change Password" />
+                </p>
+            </fieldset>
+        </div>
+    <% } %>
+</asp:Content>
