+++ samples/ASPNet/MvcTestApplication/Views/Account/ChangePasswordSuccess.aspx
@@ -1,0 +1,12 @@
+﻿<%@Language="C#" MasterPageFile="~/Views/Shared/Site.Master" Inherits="System.Web.Mvc.ViewPage" %>
+
+<asp:Content ID="changePasswordTitle" ContentPlaceHolderID="TitleContent" runat="server">
+    Change Password
+</asp:Content>
+
+<asp:Content ID="changePasswordSuccessContent" ContentPlaceHolderID="MainContent" runat="server">
+    <h2>Change Password</h2>
+    <p>
+        Your password has been changed successfully.
+    </p>
+</asp:Content>
