+++ samples/ASPNet/MvcTestApplication/Views/Account/Register.aspx
@@ -1,0 +1,60 @@
+﻿<%@ Page Language="C#" MasterPageFile="~/Views/Shared/Site.Master" Inherits="System.Web.Mvc.ViewPage<MvcTestApplication.Models.RegisterModel>" %>
+
+<asp:Content ID="registerTitle" ContentPlaceHolderID="TitleContent" runat="server">
+    Register
+</asp:Content>
+
+<asp:Content ID="registerContent" ContentPlaceHolderID="MainContent" runat="server">
+    <h2>Create a New Account</h2>
+    <p>
+        Use the form below to create a new account. 
+    </p>
+    <p>
+        Passwords are required to be a minimum of <%: ViewData["PasswordLength"] %> characters in length.
+    </p>
+
+    <% using (Html.BeginForm()) { %>
+        <%: Html.ValidationSummary(true, "Account creation was unsuccessful. Please correct the errors and try again.") %>
+        <div>
+            <fieldset>
+                <legend>Account Information</legend>
+                
+                <div class="editor-label">
+                    <%: Html.LabelFor(m => m.UserName) %>
+                </div>
+                <div class="editor-field">
+                    <%: Html.TextBoxFor(m => m.UserName) %>
+                    <%: Html.ValidationMessageFor(m => m.UserName) %>
+                </div>
+                
+                <div class="editor-label">
+                    <%: Html.LabelFor(m => m.Email) %>
+                </div>
+                <div class="editor-field">
+                    <%: Html.TextBoxFor(m => m.Email) %>
+                    <%: Html.ValidationMessageFor(m => m.Email) %>
+                </div>
+                
+                <div class="editor-label">
+                    <%: Html.LabelFor(m => m.Password) %>
+                </div>
+                <div class="editor-field">
+                    <%: Html.PasswordFor(m => m.Password) %>
+                    <%: Html.ValidationMessageFor(m => m.Password) %>
+                </div>
+                
+                <div class="editor-label">
+                    <%: Html.LabelFor(m => m.ConfirmPassword) %>
+                </div>
+                <div class="editor-field">
+                    <%: Html.PasswordFor(m => m.ConfirmPassword) %>
+                    <%: Html.ValidationMessageFor(m => m.ConfirmPassword) %>
+                </div>
+                
+                <p>
+                    <input type="submit" value="Register" />
+                </p>
+            </fieldset>
+        </div>
+    <% } %>
+</asp:Content>
