+++ samples/ASPNet/MvcTestApplication/Views/Shared/Error.aspx
@@ -1,0 +1,11 @@
+﻿<%@ Page Language="C#" MasterPageFile="~/Views/Shared/Site.Master" Inherits="System.Web.Mvc.ViewPage<System.Web.Mvc.HandleErrorInfo>" %>
+
+<asp:Content ID="errorTitle" ContentPlaceHolderID="TitleContent" runat="server">
+    Error
+</asp:Content>
+
+<asp:Content ID="errorContent" ContentPlaceHolderID="MainContent" runat="server">
+    <h2>
+        Sorry, an error occurred while processing your request.
+    </h2>
+</asp:Content>
