+++ samples/ASPNet/MvcTestApplication/Views/Shared/LogOnUserControl.ascx
@@ -1,0 +1,14 @@
+﻿<%@ Control Language="C#" Inherits="System.Web.Mvc.ViewUserControl" %>
+<%
+    if (Request.IsAuthenticated) {
+%>
+        Welcome <b><%: Page.User.Identity.Name %></b>!
+        [ <%: Html.ActionLink("Log Off", "LogOff", "Account") %> ]
+<%
+    }
+    else {
+%> 
+        [ <%: Html.ActionLink("Log On", "LogOn", "Account") %> ]
+<%
+    }
+%>
